backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
tensorflow>=2.12
Pillow>=10.0.0
numpy>=1.24.0